create index page

+++ app/views/employee/default.view.php
@@ -20,7 +20,7 @@
 </head>
 
 <div style="margin-top: 84px;">
-    <a href="employee/add" style="font-size: x-large;border: solid;border-radius: 9px;padding: 5px;margin: 151px">Add new employee</a>
+    <a href="employee/add" style="font-size: x-large;border: solid;border-radius: 9px;padding: 10px;margin: 151px;">Add new employee</a>
 </div>
 
 <table class="table table-striped table-dark" style="width: 70%; margin:auto;margin-top: 100px;">
@@ -57,8 +57,8 @@
                     <td><?= $employee->salary . " LE" ?></td>
                     <td><?= $employee->address ?></td>
                     <td>
-                        <a href="/employee/edit/"><button type="button" class="btn btn-info">Edit</button></a>
-                        <a href="/employee/delete/"><button type=" button" class="btn btn-danger">Delete</button></a>
+                        <a href="/employee/edit/<?= $employee->id ?>"><button type="button" class="btn btn-info">Edit</button></a>
+                        <a href="/employee/delete/<?= $employee->id ?>" onclick="if(!confirm('Do you want to delete this employee ?')) return false;"><i class=" fa fa-time"></i><button type=" button" class="btn btn-danger">Delete</button></a>
                     </td>
 
                 </tr>
